split files into internal and public bits
